requirements.txt
python-telegram-handler
pyTelegramBotAPI
discord-webhook
Pillow
tk
customtkinter